+++ webscout/Provider/TTS/__init__.py
@@ -11,6 +11,7 @@ from webscout.Provider.TTS.base import (
 from webscout.Provider.TTS.deepgram import DeepgramTTS
 from webscout.Provider.TTS.elevenlabs import ElevenlabsTTS
 from webscout.Provider.TTS.faster_qwen3 import FasterQwen3TTS
+from webscout.Provider.TTS.luxtts import LuxTTS
 from webscout.Provider.TTS.murfai import MurfAITTS
 from webscout.Provider.TTS.openai_fm import OpenAIFMTTS
 from webscout.Provider.TTS.parler import ParlerTTS
@@ -33,6 +34,7 @@ __all__ = [
     "DeepgramTTS",
     "ElevenlabsTTS",
     "FasterQwen3TTS",
+    "LuxTTS",
     "MurfAITTS",
     "OpenAIFMTTS",
     "ParlerTTS",